+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -13,6 +14,8 @@ class UserController extends Controller
      */
     public function index()
     {
+        Gate::authorize('read-users');
+
         return view('admin.users.index');
     }
 
@@ -21,6 +24,8 @@ class UserController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create-users');
+
         return view('admin.users.create');
     }
 
@@ -29,6 +34,8 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create-users');
+
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users',
@@ -53,6 +60,8 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        Gate::authorize('read-users');
+
         return view('admin.users.show', compact('user'));
     }
 
@@ -61,6 +70,8 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        Gate::authorize('update-users');
+
         return view('admin.users.edit', compact('user'));
     }
 
@@ -69,6 +80,8 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        Gate::authorize('update-users');
+
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users,email,' . $user->id,
@@ -97,6 +110,8 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        Gate::authorize('delete-users');
+
         $user->delete();
 
         session()->flash('swal', [
